+++ test/functional/test-3p-frame.js
@@ -166,7 +166,8 @@ describe('3p-frame', () => {
         ',"canary":true' +
         ',"hidden":false' +
         // Note that DOM fingerprint will change if the document DOM changes
-        ',"domFingerprint":"258846393"' +
+        // Note also that running it using --files uses different DOM.
+        ',"domFingerprint":"1725030182"' +
         ',"startTime":1234567888' +
         ',"amp3pSentinel":"' + amp3pSentinel + '"' +
         ',"initialIntersection":{"time":1234567888,' +
@@ -179,7 +180,6 @@ describe('3p-frame', () => {
     const srcParts = src.split('#');
     expect(srcParts[0]).to.equal(
         'http://ads.localhost:9876/dist.3p/current/frame.max.html');
-    console.log(srcParts[1]);
     expect(JSON.parse(srcParts[1])).to.deep.equal(JSON.parse(fragment));
 
     // Switch to same origin for inner tests.
